refactor: extract logic from CreateEditModal component in hooks and utils
